Placeholder Namespace + name privacy
