+++ components/ILIAS/Mail/classes/Signature/MailInstallationSignature.php
@@ -21,6 +21,10 @@ declare(strict_types=1);
 namespace ILIAS\Mail\Signature;
 
 use ilSetting;
+use ILIAS\Mail\Placeholder\MailSignatureIliasUrlPlaceholder;
+use ILIAS\Mail\Placeholder\MailSignatureInstallationNamePlaceholder;
+use ILIAS\Mail\Placeholder\MailSignatureInstallationDescriptionPlaceholder;
+use ILIAS\Mail\Placeholder\Placeholder;
 
 class MailInstallationSignature implements Signature
 {
